Show how you could use matrix-nio to assert or generate messages.
