+++ setup.cfg
@@ -17,7 +17,7 @@ install_requires =
   attrs
   aiohttp
   quart
-  matrix-nio
+  matrix-nio[e2e]
 
 [options.package_data]
 trafficlight = py.typed
